add pagebackgroundcontext and use it on image galleries
